Added TAP for XG hotend & BambuLab hotend as new one
I have created and attached the TAP mount for the Phaetus XG hotend. I also added the BambuLab hotend and created a mount for TAP. Faceplate I have put the holes to the top, so that the screws supplied by BambuLab can still be used on the fan. For mounting the hotends top 2x M3x40 screws are needed and 2x heatset inserts M3